Refactored Frontend
